+++ test/reactionsystem_core/reactionsystem.jl
@@ -503,6 +503,7 @@ let
         @named rs = ReactionSystem(eqs, t)
         rs = complete(rs)
         ssys = complete(make_cle_sde(rs))
+        ssys = MT.mtkcompile(ssys)  # Required for Brownian-based SDE systems with constraints
         @test issetequal(MT.get_unknowns(ssys), [B, C, D, E])
         _ps = filter(!isinitial, MT.get_ps(ssys))
         @test issetequal(_ps, [A, k1, k2])
@@ -1089,7 +1090,7 @@ let
         cevents = [[V ~ 2.0] => [V ~ V/2, A ~ A/2]]
         @named rs = ReactionSystem(vcat(rxs, eqs), t; continuous_events = cevents)
         rs = complete(rs)
-        sys = complete(make_sck_jump(rs))
+        sys = complete(make_hybrid_model(rs; default_scale = PhysicalScale.Jump))
         @test sys isa MT.System
         @test MT.has_equations(sys)
         @test length(massactionjumps(sys)) == 1
@@ -1115,7 +1116,7 @@ let
         cevents = [[V ~ 2.0] => [V ~ V/2, A ~ A/2]]
         @named rs = ReactionSystem(vcat(rxs, eqs), t; continuous_events = cevents)
         rs = complete(rs)
-        sys = complete(make_sck_jump(rs))
+        sys = complete(make_hybrid_model(rs; default_scale = PhysicalScale.Jump))
         @test sys isa MT.System
         @test MT.has_equations(sys)
         @test isempty(massactionjumps(sys))